+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v0.7.2
 
 1. Migrated to uhppoted-lib `config` command implementation
+2. Reworked `get-events` to return `first`, `last` and `current` event indices.
 
 ### v0.7.1